Matrix: isolate credential write runtime
(cherry picked from commit f4f0b171d3bcdfd88f051e1d2f8b852ff1f0eafa)
