+++ extensions/matrix/src/matrix/client/config.ts
@@ -6,13 +6,10 @@ import { resolveMatrixAccountStringValues } from "../../auth-precedence.js";
 import { getMatrixScopedEnvVarNames } from "../../env-vars.js";
 import {
   DEFAULT_ACCOUNT_ID,
-  assertHttpUrlTargetsPrivateNetwork,
   isPrivateOrLoopbackHost,
-  type LookupFn,
   normalizeAccountId,
   normalizeOptionalAccountId,
   normalizeResolvedSecretInputString,
-  ssrfPolicyFromAllowPrivateNetwork,
 } from "../../runtime-api.js";
 import { getMatrixRuntime } from "../../runtime.js";
 import type { CoreConfig } from "../../types.js";
@@ -70,21 +67,6 @@ function readMatrixAccountConfigField(
 
 function clampMatrixInitialSyncLimit(value: unknown): number | undefined {
   return typeof value === "number" ? Math.max(0, Math.floor(value)) : undefined;
-}
-
-const MATRIX_HTTP_HOMESERVER_ERROR =
-  "Matrix homeserver must use https:// unless it targets a private or loopback host";
-
-function buildMatrixNetworkFields(
-  allowPrivateNetwork: boolean | undefined,
-): Pick<MatrixResolvedConfig, "allowPrivateNetwork" | "ssrfPolicy"> {
-  if (!allowPrivateNetwork) {
-    return {};
-  }
-  return {
-    allowPrivateNetwork: true,
-    ssrfPolicy: ssrfPolicyFromAllowPrivateNetwork(true),
-  };
 }
 
 function resolveGlobalMatrixEnvConfig(env: NodeJS.ProcessEnv): MatrixEnvConfig {
@@ -181,10 +163,7 @@ export function hasReadyMatrixEnvAuth(config: {
   return Boolean(homeserver && (accessToken || (userId && password)));
 }
 
-export function validateMatrixHomeserverUrl(
-  homeserver: string,
-  opts?: { allowPrivateNetwork?: boolean },
-): string {
+export function validateMatrixHomeserverUrl(homeserver: string): string {
   const trimmed = clean(homeserver, "matrix.homeserver");
   if (!trimmed) {
     throw new Error("Matrix homeserver is required (matrix.homeserver)");
@@ -209,28 +188,13 @@ export function validateMatrixHomeserverUrl(
   if (parsed.search || parsed.hash) {
     throw new Error("Matrix homeserver URL must not include query strings or fragments");
   }
-  if (
-    parsed.protocol === "http:" &&
-    opts?.allowPrivateNetwork !== true &&
-    !isPrivateOrLoopbackHost(parsed.hostname)
-  ) {
-    throw new Error(MATRIX_HTTP_HOMESERVER_ERROR);
+  if (parsed.protocol === "http:" && !isPrivateOrLoopbackHost(parsed.hostname)) {
+    throw new Error(
+      "Matrix homeserver must use https:// unless it targets a private or loopback host",
+    );
   }
 
   return trimmed;
-}
-
-export async function resolveValidatedMatrixHomeserverUrl(
-  homeserver: string,
-  opts?: { allowPrivateNetwork?: boolean; lookupFn?: LookupFn },
-): Promise<string> {
-  const normalized = validateMatrixHomeserverUrl(homeserver, opts);
-  await assertHttpUrlTargetsPrivateNetwork(normalized, {
-    allowPrivateNetwork: opts?.allowPrivateNetwork,
-    lookupFn: opts?.lookupFn,
-    errorMessage: MATRIX_HTTP_HOMESERVER_ERROR,
-  });
-  return normalized;
 }
 
 export function resolveMatrixConfig(
@@ -255,7 +219,6 @@ export function resolveMatrixConfig(
   });
   const initialSyncLimit = clampMatrixInitialSyncLimit(matrix.initialSyncLimit);
   const encryption = matrix.encryption ?? false;
-  const allowPrivateNetwork = matrix.allowPrivateNetwork === true ? true : undefined;
   return {
     homeserver: resolvedStrings.homeserver,
     userId: resolvedStrings.userId,
@@ -265,7 +228,6 @@ export function resolveMatrixConfig(
     deviceName: resolvedStrings.deviceName || undefined,
     initialSyncLimit,
     encryption,
-    ...buildMatrixNetworkFields(allowPrivateNetwork),
   };
 }
 
@@ -308,8 +270,6 @@ export function resolveMatrixConfigForAccount(
     accountInitialSyncLimit ?? clampMatrixInitialSyncLimit(matrix.initialSyncLimit);
   const encryption =
     typeof account.encryption === "boolean" ? account.encryption : (matrix.encryption ?? false);
-  const allowPrivateNetwork =
-    account.allowPrivateNetwork === true || matrix.allowPrivateNetwork === true ? true : undefined;
 
   return {
     homeserver: resolvedStrings.homeserver,
@@ -320,7 +280,6 @@ export function resolveMatrixConfigForAccount(
     deviceName: resolvedStrings.deviceName || undefined,
     initialSyncLimit,
     encryption,
-    ...buildMatrixNetworkFields(allowPrivateNetwork),
   };
 }
 
@@ -379,9 +338,7 @@ export async function resolveMatrixAuth(params?: {
   accountId?: string | null;
 }): Promise<MatrixAuth> {
   const { cfg, env, accountId, resolved } = resolveMatrixAuthContext(params);
-  const homeserver = await resolveValidatedMatrixHomeserverUrl(resolved.homeserver, {
-    allowPrivateNetwork: resolved.allowPrivateNetwork,
-  });
+  const homeserver = validateMatrixHomeserverUrl(resolved.homeserver);
   let credentialsWriter: typeof import("../credentials-write.runtime.js") | undefined;
   const loadCredentialsWriter = async () => {
     credentialsWriter ??= await import("../credentials-write.runtime.js");
@@ -410,9 +367,7 @@ export async function resolveMatrixAuth(params?: {
     if (!userId || !knownDeviceId) {
       // Fetch whoami when we need to resolve userId and/or deviceId from token auth.
       ensureMatrixSdkLoggingConfigured();
-      const tempClient = new MatrixClient(homeserver, resolved.accessToken, undefined, undefined, {
-        ssrfPolicy: resolved.ssrfPolicy,
-      });
+      const tempClient = new MatrixClient(homeserver, resolved.accessToken);
       const whoami = (await tempClient.doRequest("GET", "/_matrix/client/v3/account/whoami")) as {
         user_id?: string;
         device_id?: string;
@@ -460,7 +415,6 @@ export async function resolveMatrixAuth(params?: {
       deviceName: resolved.deviceName,
       initialSyncLimit: resolved.initialSyncLimit,
       encryption: resolved.encryption,
-      ...buildMatrixNetworkFields(resolved.allowPrivateNetwork),
     };
   }
 
@@ -477,7 +431,6 @@ export async function resolveMatrixAuth(params?: {
       deviceName: resolved.deviceName,
       initialSyncLimit: resolved.initialSyncLimit,
       encryption: resolved.encryption,
-      ...buildMatrixNetworkFields(resolved.allowPrivateNetwork),
     };
   }
 
@@ -493,9 +446,7 @@ export async function resolveMatrixAuth(params?: {
 
   // Login with password using the same hardened request path as other Matrix HTTP calls.
   ensureMatrixSdkLoggingConfigured();
-  const loginClient = new MatrixClient(homeserver, "", undefined, undefined, {
-    ssrfPolicy: resolved.ssrfPolicy,
-  });
+  const loginClient = new MatrixClient(homeserver, "");
   const login = (await loginClient.doRequest("POST", "/_matrix/client/v3/login", undefined, {
     type: "m.login.password",
     identifier: { type: "m.id.user", user: resolved.userId },
@@ -523,7 +474,6 @@ export async function resolveMatrixAuth(params?: {
     deviceName: resolved.deviceName,
     initialSyncLimit: resolved.initialSyncLimit,
     encryption: resolved.encryption,
-    ...buildMatrixNetworkFields(resolved.allowPrivateNetwork),
   };
 
   const { saveMatrixCredentials } = await loadCredentialsWriter();
